Add files via upload
This Project dives deep into the Adidas Sales Dataset for the US consumer base for the years 2020-2021. This time period saw some of the most interesting macroeconomic events in the US economy including the effect of the COVID-19 Pandemic on the US consumer behavior, massive changes in the unemployment figures, supply chain constraints, and a post-pandemic boom. Using SQL, we try to analyze the nationwide sales data for different Adidas products. We also furthered our analysis by analyzing regional, statewide, and citywide behaviors, all of which had different behaviors from the happenings of this period.